[exec] incorporating with_klio_config into exec run
